Move tests to two main sub folders unit and integration.
