+++ src/test/java/com/linkvault/integration/link/LinkControllerIntegrationTest.java
@@ -1,0 +1,21 @@
+package com.linkvault.integration.link;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.web.servlet.MockMvc;
+
+@ActiveProfiles("test")
+@SpringBootTest
+@AutoConfigureMockMvc
+@TestPropertySource("classpath:application-test.properties")
+public class LinkControllerIntegrationTest {
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Value("${JWT_SECRET}")
+    private String jwtSecret;
+}
