+++ src/test/java/com/linkvault/unit/security/JwtUtilsTest.java
@@ -1,4 +1,4 @@
-package com.linkvault.security;
+package com.linkvault.unit.security;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
